+++ src/lcd/lcd.h
@@ -51,8 +51,8 @@ int lcdGetDescaleCycle(void);
 
 void lcdSetDescaleCycle(int cycle);
 void lcdSetPressure(float val);
-void lcdSetTemperature(uint16_t val);
-void lcdSetTemperatureDecimal(uint16_t val);
+void lcdSetTemperature(float val);
+void lcdSetTemperatureDecimal(float val);
 void lcdTargetState(int val);
 void lcdSetWeight(float val);
 void lcdSetFlow(int val);
